Prevent excessive error logs when deleting file caches.
The glob will match a directory, boom - errors on unlink. Gross.

+++ src/sprout/Helpers/File.php
@@ -20,6 +20,7 @@ use Kohana;
 
 use karmabunny\pdb\Exceptions\RowMissingException;
 
+use Throwable;
 
 /**
  * Methods for working with files, including images
@@ -386,9 +387,11 @@ class File
         // Old cache structure
         $files = glob(STORAGE_PATH . 'cache/resize-*-' . $filename);
         foreach ($files as $file) {
+            if (is_dir($file)) continue;
+
             try {
                 unlink($file);
-            } catch (Exception $ex) {
+            } catch (Throwable $ex) {
                 // Log and continue
                 Kohana::logException($ex);
             }
@@ -397,9 +400,11 @@ class File
         // Updated cache structure
         $files = glob(WEBROOT . "files/resize/*/{$filename}");
         foreach ($files as $file) {
+            if (is_dir($file)) continue;
+
             try {
                 unlink($file);
-            } catch (Exception $ex) {
+            } catch (Throwable $ex) {
                 // Log and continue
                 Kohana::logException($ex);
             }
